Increase the length of the Dungeon to 18 levels and the Depths to 9 levels
This keeps the main trunk of the game as long as it was prior to the
late Dungeon being split off into the Depths. I dislike the upstream
tendency to shorten the game (see also: Elven Halls, Vaults). This may
mess with the intended experience gain curves, but that can be tweaked
in other ways if necessary.