Reduce overall XP value of monsters by 10%
This is intended to compensate for the general lengthening of the game
in this fork. There are more levels in Dungeon, Depths, Elf, Dwarf, and
Vaults relative to upstream. I'd like this to make the game longer, but
not significantly easier, especially when the additional experience from
those extra levels has accumulated by the end of the game.